Remove the tree-walk interpreter
Having it was kind of fun and I had plans for it (like using it for
evaluating constant expressions at compile-time for optimization
purposes). However, I've found myself (over and over again) complicating
the implementation to allow the VM and tree-walk interpreters to
coexist. It's getting especially tricky for the testing infrastructure
to work with both.

So, I cannot justify keeping the tree-walk interpreter anymore. I'll
find a way to use the VM for compile-time execution (which shall be the
right thing to do anyway, given it's supposed to be considerably
faster).

Farewell, TWI!