Redesign versioning
This new design is much simpler, it doesn't try to do anything I'd
expect from a "normal" programming language. Specifically, it doesn't
try to keep the relationship between versions of source code and binary
code. If you want to have access to the source code that generated a
given version of the binary, just use version control!